+++ example/react-typescript4.1/namespaces/src/i18n/config.ts
@@ -1,0 +1,17 @@
+import i18n from 'i18next';
+import ns1 from './en/ns1.json';
+import ns2 from './en/ns2.json';
+import { initReactI18next } from 'react-i18next';
+
+export const resources = {
+  en: {
+    ns1,
+    ns2,
+  },
+} as const;
+
+i18n.use(initReactI18next).init({
+  lng: 'en',
+  ns: ['ns1', 'ns2'],
+  resources,
+});